requirements.txt
Django==4.2.27
djangorestframework==3.16.1
autoflake==1.4
black==22.3.0
pylint==2.14.5
django-cors-headers==4.9.0
psycopg2==2.9.9